use numpy C-API directly
