+++ py-polars/polars/dataframe/frame.py
@@ -41,7 +41,6 @@ from polars.datatypes import (
     Float64,
     Object,
     String,
-    dtype_to_ctype,
     py_type_to_dtype,
 )
 from polars.dependencies import (
@@ -2163,14 +2162,9 @@ class DataFrame:
                 out[c] = arrays[idx]
         else:
             if order == "fortran":
-                ptr = self._df.to_numpy_view()
-                if ptr is not None:
-                    from polars.series._numpy import SeriesView, _ptr_to_numpy
-
-                    ptr_type = dtype_to_ctype(self[:, 0].dtype)
-                    array = _ptr_to_numpy(ptr, (self.height, self.width), ptr_type)
-                    array.setflags(write=False)
-                    return np.array(SeriesView(array, self), copy=False, order="F")
+                array = self._df.to_numpy_view()
+                if array is not None:
+                    return array
 
             out = self._df.to_numpy(order)
             if out is None:
