+++ py-polars/src/to_numpy.rs
@@ -1,0 +1,127 @@
+use ndarray::IntoDimension;
+use numpy::npyffi::{flags, PyArrayObject};
+use numpy::{npyffi, Element, IntoPyArray, ToNpyDims, PY_ARRAY_API};
+use polars_core::prelude::*;
+use polars_core::utils::try_get_supertype;
+use polars_core::with_match_physical_numeric_polars_type;
+use pyo3::prelude::*;
+use pyo3::{IntoPy, PyAny, PyObject, Python};
+
+use crate::conversion::Wrap;
+use crate::dataframe::PyDataFrame;
+
+#[pymethods]
+#[allow(clippy::wrong_self_convention)]
+impl PyDataFrame {
+    pub fn to_numpy_view(&self, py: Python) -> Option<PyObject> {
+        if self.df.is_empty() {
+            return None;
+        }
+        let first = self.df.get_columns().first().unwrap().dtype();
+        if !first.is_numeric() {
+            return None;
+        }
+        if !self
+            .df
+            .get_columns()
+            .iter()
+            .all(|s| s.null_count() == 0 && s.dtype() == first && s.chunks().len() == 1)
+        {
+            return None;
+        }
+
+        // Object to the dataframe keep the memory alive.
+        let owner = self.clone().into_py(py);
+
+        fn get_ptr<T: PolarsNumericType>(
+            py: Python,
+            columns: &[Series],
+            owner: PyObject,
+        ) -> Option<PyObject>
+        where
+            T::Native: Element,
+        {
+            let slices = columns
+                .iter()
+                .map(|s| {
+                    let ca: &ChunkedArray<T> = s.unpack().unwrap();
+                    ca.cont_slice().unwrap()
+                })
+                .collect::<Vec<_>>();
+
+            let first = slices.first().unwrap();
+            unsafe {
+                let mut end_ptr = first.as_ptr().add(first.len());
+                // Check if all arrays are from the same buffer
+                let all_contiguous = slices[1..].iter().all(|slice| {
+                    let valid = slice.as_ptr() == end_ptr;
+
+                    end_ptr = slice.as_ptr().add(slice.len());
+
+                    valid
+                });
+
+                if all_contiguous {
+                    let start_ptr = first.as_ptr();
+
+                    let mut dims = [first.len(), columns.len()].into_dimension();
+
+                    // See: https://numpy.org/doc/stable/reference/c-api/array.html
+                    let array = PY_ARRAY_API.PyArray_NewFromDescr(
+                        py,
+                        PY_ARRAY_API.get_type_object(py, npyffi::NpyTypes::PyArray_Type),
+                        T::Native::get_dtype(py).into_dtype_ptr(),
+                        dims.ndim_cint(),
+                        dims.as_dims_ptr(),
+                        std::ptr::null_mut(),
+                        start_ptr as _,              // data
+                        flags::NPY_ARRAY_OUT_FARRAY, // flag
+                        std::ptr::null_mut(),        //obj
+                    );
+
+                    // This keeps the memory alive
+                    let owner_ptr = owner.as_ptr();
+                    PY_ARRAY_API.PyArray_SetBaseObject(py, array as *mut PyArrayObject, owner_ptr);
+
+                    let any: &PyAny = py.from_owned_ptr(array);
+                    Some(any.into_py(py))
+                } else {
+                    None
+                }
+            }
+        }
+        with_match_physical_numeric_polars_type!(first, |$T| {
+            get_ptr::<$T>(py, self.df.get_columns(), owner)
+        })
+    }
+
+    pub fn to_numpy(&self, py: Python, order: Wrap<IndexOrder>) -> Option<PyObject> {
+        let mut st = None;
+        for s in self.df.iter() {
+            let dt_i = s.dtype();
+            match st {
+                None => st = Some(dt_i.clone()),
+                Some(ref mut st) => {
+                    *st = try_get_supertype(st, dt_i).ok()?;
+                },
+            }
+        }
+        let st = st?;
+
+        #[rustfmt::skip]
+            let pyarray = match st {
+            DataType::UInt8 => self.df.to_ndarray::<UInt8Type>(order.0).ok()?.into_pyarray(py).into_py(py),
+            DataType::Int8 => self.df.to_ndarray::<Int8Type>(order.0).ok()?.into_pyarray(py).into_py(py),
+            DataType::UInt16 => self.df.to_ndarray::<UInt16Type>(order.0).ok()?.into_pyarray(py).into_py(py),
+            DataType::Int16 => self.df.to_ndarray::<Int16Type>(order.0).ok()?.into_pyarray(py).into_py(py),
+            DataType::UInt32 => self.df.to_ndarray::<UInt32Type>(order.0).ok()?.into_pyarray(py).into_py(py),
+            DataType::UInt64 => self.df.to_ndarray::<UInt64Type>(order.0).ok()?.into_pyarray(py).into_py(py),
+            DataType::Int32 => self.df.to_ndarray::<Int32Type>(order.0).ok()?.into_pyarray(py).into_py(py),
+            DataType::Int64 => self.df.to_ndarray::<Int64Type>(order.0).ok()?.into_pyarray(py).into_py(py),
+            DataType::Float32 => self.df.to_ndarray::<Float32Type>(order.0).ok()?.into_pyarray(py).into_py(py),
+            DataType::Float64 => self.df.to_ndarray::<Float64Type>(order.0).ok()?.into_pyarray(py).into_py(py),
+            _ => return None,
+        };
+        Some(pyarray)
+    }
+}
